[SET-626] Split large comments into chunks
Salesforce restricts the size of a comment to 4000 charactes. This
change adds a function that splits comments into chunks no larger than
4000 characters. The chunks are posted in separate comments. The chunk
size is configurable to allow for future changes in how Salesforce
restricts comments.

In order to support default values in the configuration files so that
the current comment length does not have to be set explicitly, refactor
the configuration related data types.  Since Go does not support default
values within the `struct` definition such defaults have to be supplied
via a factory function.

Closes: https://github.com/canonical/athena-core/issues/121
Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>